created styles.js
