--- src/components/PrettyButton.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import Button from 'react-bootstrap/Button';
-
-function PrettyButton() {
-    <Button>
-        Learn More
-    </Button>
-}
-
-export default PrettyButton;
