--- src/components/PrettyCard.css
@@ -1,20 +1,0 @@
-.PrettyCard {
-    margin: 50px;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    background-color: #ffcbce;
-    transition: transform 0.3s ease-in-out;
-    padding: 50px;
-}
-
-.PrettyCard:hover {
-    transform: scale(1.2);   
-}
-
-img {
-    width: 200px;
-    height: 200px;
-    align-self: center;
-}
-
