+++ src/components/Timeline.js
@@ -6,6 +6,7 @@ import StepLabel from '@mui/material/StepLabel';
 import StepContent from '@mui/material/StepContent';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
+import stepperStyles from './Styles';
 
 function Timeline({steps}) {
   const [activeStep, setActiveStep] = React.useState(0);
@@ -18,49 +19,12 @@ function Timeline({steps}) {
     setActiveStep((prevActiveStep) => prevActiveStep - 1);
   };
 
-  const stepperStyles = {
-    '& .MuiStepContent-root': { // THIS ENTIRE SECTION NEEDS TO BE REFACTORED!!!
-      borderLeft: '1px solid black',
-    },
-    '& .MuiStepConnector-line': {
-      borderLeft: '1px solid black',
-    },
-    '& .MuiButton-containedPrimary': {
-      backgroundColor: 'black',
-      border: '1px',
-      '&:hover': {
-        backgroundColor: '#3c434a', 
-      },
-    },
-    '& .MuiButton-textPrimary': {
-      color: 'black',
-      border: '1px',
-    },
-    '& .MuiStepIcon-root.Mui-active': {
-      color: 'black', 
-    },
-    '& .MuiStepIcon-root.Mui-completed': {
-      color: 'black', 
-    },
-    '& .MuiSvgIcon-root': {
-      borderRadius: '1px',
-      color: 'black', 
-    },
-  }
-
   return (
     <Box> 
       <Stepper activeStep={activeStep} orientation="vertical" sx={stepperStyles}>
         {steps.map((step, index) => (
           <Step key={step.label}>
-            <StepLabel
-              sx={stepperStyles}
-              optional={
-                index === steps.length - 1 ? (
-                  <Typography variant="caption">Last step</Typography>
-                ) : null
-              }
-            >
+            <StepLabel sx={stepperStyles}>
               {step.label}
             </StepLabel>
             <StepContent>
